Refactor auto-consumption
Code cleanup: Added a new record type containing (food item, space,
adventures) instead of passing around an array for each field.

The new record type ConsumeAction (name pending) stores a bunch of
useful information like "how are we gonna get this?" and "how many
adventures should we pretend this gives when deciding whether to consume
it?" The first is useful for reasoning about pulls (and in the future,
time-spinning/diet pills?), and the second is useful for (things like)
prioritizing pulling and eating Key Lime Pies in softcore when we don't
yet have enough tower keys.